+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 Fixed updating debugger settings when its configuration file changed.
 - [#30](https://github.com/denis-shienkov/vscode-qbs/issues/30)
 Implemented ability to select the 'Auto' entry for the debugging.
+- [#4](https://github.com/denis-shienkov/vscode-qbs/issues/4)
+Right now the pre-defined compiler macros also reported for the
+non-generic compilers (like KEIL, IAR, SDCC).
 
 ## 0.0.8